fix: refactor create config implementation function
