+++ src/implementations/config/create.ts
@@ -1,38 +1,13 @@
-import {
-  Issuer,
-  RevocationType,
-  IdentityProofType,
-  OpenAttestationDocument,
-} from "@govtechsg/open-attestation/dist/types/__generated__/schema.2.0";
+import { Issuer } from "@govtechsg/open-attestation/dist/types/__generated__/schema.2.0";
 import fs from "fs";
 import path from "path";
 import { info } from "signale";
 import { readFile } from "../../implementations/utils/disk";
-import tradetrustConfig from "./templates/tradetrust.json";
 import { handler as createTemporaryDns } from "../../commands/dns/txt-record/create";
 import { CreateConfigCommand } from "../../commands/config/config.type";
-import { getDocumentStoreAddress } from "./documentStore";
-import { getTokenRegistryAddress } from "./tokenRegistry";
-
-interface ConfigFile {
-  wallet: EncryptedWallet;
-  forms: Form[];
-}
-
-interface EncryptedWallet {
-  type: "ENCRYPTED_JSON";
-  encryptedJson: string;
-}
-
-interface Form {
-  type: "VERIFIABLE_DOCUMENT" | "TRANSFERABLE_RECORD";
-  defaults: OpenAttestationDocument;
-}
-
-interface TypesOfForms {
-  type: "VERIFIABLE_DOCUMENT" | "TRANSFERABLE_RECORD";
-  identityProofTypes: (IdentityProofType | undefined)[];
-}
+import { ConfigFile, TypesOfForms } from "./types";
+import { Wallet } from "ethers";
+import { getDocumentStoreOrTokenRegistryAddress, updateConfigFile, updateForms, getConfigFile } from "./helpers";
 
 const sandboxEndpointUrl = "https://sandbox.fyntech.io";
 
@@ -42,151 +17,82 @@ export const create = async ({
   configType,
   configTemplatePath,
 }: CreateConfigCommand): Promise<string | undefined> => {
-  try {
-    const wallet = await readFile(encryptedWalletPath);
-    const walletObject = JSON.parse(wallet);
+  const wallet = await readFile(encryptedWalletPath);
+  const walletObject = JSON.parse(wallet) as Wallet;
 
-    info(`Wallet detected at ${encryptedWalletPath}`);
+  info(`Wallet detected at ${encryptedWalletPath}`);
 
-    const configFile: ConfigFile = await getConfigFile(configType, configTemplatePath);
+  const configFile: ConfigFile = await getConfigFile(configType, configTemplatePath);
 
-    const { forms } = configFile;
+  const { forms } = configFile;
 
-    const getContractAddress = async (
-      typesOfForms: TypesOfForms[],
-      formType: "VERIFIABLE_DOCUMENT" | "TRANSFERABLE_RECORD",
-      identityProofType: "DNS-DID" | "DNS-TXT"
-    ): Promise<string> => {
-      const isValidForm = typesOfForms.some(
-        (item: TypesOfForms) =>
-          item.type === formType && (<any>Object).values(item.identityProofTypes).includes(identityProofType)
-      );
-
-      if (isValidForm) {
-        switch (true) {
-          case formType === "VERIFIABLE_DOCUMENT" && identityProofType === "DNS-TXT":
-            return await getDocumentStoreAddress(encryptedWalletPath);
-
-          case formType === "VERIFIABLE_DOCUMENT" && identityProofType === "DNS-DID":
-            info(`Creating temporary DNS for DID`);
-            return (
-              (await createTemporaryDns({
-                networkId: 3,
-                publicKey: `did:ethr:0x${walletObject.address}#controller`,
-                sandboxEndpoint: sandboxEndpointUrl,
-              })) || ""
-            );
-
-          case formType === "TRANSFERABLE_RECORD":
-            return await getTokenRegistryAddress(encryptedWalletPath);
-
-          default:
-            throw new Error("Invalid form detected in config file, please update the form before proceeding.");
-        }
-      } else {
-        throw new Error("Invalid form detected in config file, please update the form before proceeding.");
-      }
-    };
-
-    // loop through the form template to check the type of forms
-    const typesOfForms: TypesOfForms[] = forms.map((form) => {
-      const identityProofTypes = form.defaults.issuers.map((issuer: Issuer) => issuer.identityProof?.type);
-      return {
-        type: form.type,
-        identityProofTypes: identityProofTypes,
-      };
-    });
-
-    // generate doc store, token registry and DNS based on the form type in the form template
-    const documentStoreAddress = await getContractAddress(typesOfForms, "VERIFIABLE_DOCUMENT", "DNS-TXT");
-    const verifiableDocumentDnsTxtName = documentStoreAddress
-      ? await createTemporaryDns({ networkId: 3, address: documentStoreAddress, sandboxEndpoint: sandboxEndpointUrl })
-      : "";
-    const verifiableDocumentDnsDidName = await getContractAddress(typesOfForms, "VERIFIABLE_DOCUMENT", "DNS-DID");
-    const tokenRegistryAddress = await getContractAddress(typesOfForms, "TRANSFERABLE_RECORD", "DNS-TXT");
-    const tokenRegistryDnsName = tokenRegistryAddress
-      ? await createTemporaryDns({ networkId: 3, address: tokenRegistryAddress, sandboxEndpoint: sandboxEndpointUrl })
-      : "";
-
-    const updatedForms = updateForms(
-      forms,
-      documentStoreAddress,
-      tokenRegistryAddress,
-      walletObject,
-      verifiableDocumentDnsTxtName || "",
-      verifiableDocumentDnsDidName,
-      tokenRegistryDnsName || ""
+  const getContractAddress = async (
+    typesOfForms: TypesOfForms[],
+    formType: "VERIFIABLE_DOCUMENT" | "TRANSFERABLE_RECORD",
+    identityProofType: "DNS-DID" | "DNS-TXT"
+  ): Promise<string> => {
+    const isValidForm = typesOfForms.some(
+      (item: TypesOfForms) =>
+        item.type === formType && (<any>Object).values(item.identityProofTypes).includes(identityProofType)
     );
 
-    const updatedConfigFile = updateConfigFile(configFile, wallet, updatedForms);
+    if (!isValidForm)
+      throw new Error("Invalid form detected in config file, please update the form before proceeding.");
 
-    const configFileName = "config.json";
-    const outputPath = path.join(outputDir, configFileName);
-    fs.writeFileSync(outputPath, JSON.stringify(updatedConfigFile, null, 2));
-    return outputPath;
-  } catch (e) {
-    if (e instanceof Error) {
-      throw new Error(e.message);
+    switch (true) {
+      case formType === "TRANSFERABLE_RECORD":
+      case formType === "VERIFIABLE_DOCUMENT" && identityProofType === "DNS-TXT":
+        return await getDocumentStoreOrTokenRegistryAddress(encryptedWalletPath, formType);
+
+      case formType === "VERIFIABLE_DOCUMENT" && identityProofType === "DNS-DID":
+        info(`Creating temporary DNS for DID`);
+        return (
+          (await createTemporaryDns({
+            networkId: 3,
+            publicKey: `did:ethr:0x${walletObject.address}#controller`,
+            sandboxEndpoint: sandboxEndpointUrl,
+          })) || ""
+        );
+
+      default:
+        throw new Error("Invalid form detected in config file, please update the form before proceeding.");
     }
-  }
-};
+  };
 
-const getConfigFile = async (configType: string, configTemplatePath: string): Promise<ConfigFile> => {
-  switch (configType) {
-    case "tradetrust":
-      return tradetrustConfig as ConfigFile;
-
-    default:
-      return JSON.parse(await readFile(configTemplatePath));
-  }
-};
-
-const updateForms = (
-  forms: Form[],
-  documentStoreAddress: string,
-  tokenRegistryAddress: string,
-  walletObject: any,
-  verifiableDocumentDnsTxtName: string,
-  verifiableDocumentDnsDidName: string,
-  tokenRegistryDnsName: string
-): Form[] => {
-  // replace the values in the forms with the updated value
-  const updatedForms = forms;
-  updatedForms.forEach((form) => {
-    if (form.type === "VERIFIABLE_DOCUMENT") {
-      const updatedIssuers = form.defaults.issuers.map((issuer: Issuer) => {
-        if (issuer.identityProof?.type === "DNS-TXT") {
-          issuer.name = "DEMO DOCUMENT STORE";
-          issuer.documentStore = documentStoreAddress;
-          if (issuer.identityProof?.location) issuer.identityProof.location = verifiableDocumentDnsTxtName;
-        } else if (issuer.identityProof?.type === "DNS-DID") {
-          issuer.name = "DEMO ISSUER";
-          issuer.id = `did:ethr:0x${walletObject.address}`;
-          if (issuer.revocation?.type) issuer.revocation.type = "NONE" as RevocationType;
-          if (issuer.identityProof?.location) issuer.identityProof.location = verifiableDocumentDnsDidName;
-          if (issuer.identityProof?.key) issuer.identityProof.key = `did:ethr:0x${walletObject.address}#controller`;
-        }
-        return issuer;
-      });
-      form.defaults.issuers = updatedIssuers;
-    }
-    if (form.type === "TRANSFERABLE_RECORD") {
-      const updatedIssuers = form.defaults.issuers.map((issuer: Issuer) => {
-        issuer.name = "DEMO TOKEN REGISTRY";
-        issuer.tokenRegistry = tokenRegistryAddress;
-        if (issuer.identityProof?.location) issuer.identityProof.location = tokenRegistryDnsName;
-        return issuer;
-      });
-      form.defaults.issuers = updatedIssuers;
-    }
+  // loop through the form template to check the type of forms
+  const typesOfForms: TypesOfForms[] = forms.map((form) => {
+    const identityProofTypes = form.defaults.issuers.map((issuer: Issuer) => issuer.identityProof?.type);
+    return {
+      type: form.type,
+      identityProofTypes: identityProofTypes,
+    };
   });
-  return updatedForms;
-};
 
-const updateConfigFile = (configFile: ConfigFile, wallet: string, updatedForms: Form[]): ConfigFile => {
-  const updatedConfigFile = configFile;
-  updatedConfigFile.wallet.encryptedJson = wallet;
+  // generate doc store, token registry and DNS based on the form type in the form template
+  const documentStoreAddress = await getContractAddress(typesOfForms, "VERIFIABLE_DOCUMENT", "DNS-TXT");
+  const verifiableDocumentDnsTxtName = documentStoreAddress
+    ? await createTemporaryDns({ networkId: 3, address: documentStoreAddress, sandboxEndpoint: sandboxEndpointUrl })
+    : "";
+  const verifiableDocumentDnsDidName = await getContractAddress(typesOfForms, "VERIFIABLE_DOCUMENT", "DNS-DID");
+  const tokenRegistryAddress = await getContractAddress(typesOfForms, "TRANSFERABLE_RECORD", "DNS-TXT");
+  const tokenRegistryDnsName = tokenRegistryAddress
+    ? await createTemporaryDns({ networkId: 3, address: tokenRegistryAddress, sandboxEndpoint: sandboxEndpointUrl })
+    : "";
 
-  updatedConfigFile.forms = updatedForms;
-  return updatedConfigFile;
+  const updatedForms = updateForms(
+    forms,
+    documentStoreAddress,
+    tokenRegistryAddress,
+    walletObject,
+    verifiableDocumentDnsTxtName || "",
+    verifiableDocumentDnsDidName,
+    tokenRegistryDnsName || ""
+  );
+
+  const updatedConfigFile = updateConfigFile(configFile, wallet, updatedForms);
+
+  const configFileName = "config.json";
+  const outputPath = path.join(outputDir, configFileName);
+  fs.writeFileSync(outputPath, JSON.stringify(updatedConfigFile, null, 2));
+  return outputPath;
 };
